Reworked Depot operations
- Renamed "Materials Generator" dialogue to "Depot Generator" and reworked UI
- Moved Depot Generator from Game to Tools
- Set Depot Generator to dialogue window and fixed awkward behavior when setting Depot Path
- Added "Open Game Folder" to Game
- Moved "Unbundle Game" from menu bar to Depot Generator
- Set MLMASK to export with "AsList" argument
- Added log messages for when Depot operations finish
